requirements.txt
streamlit==1.31.0
langchain==0.1.5
langchain-experimental==0.0.49
google-generativeai==0.3.2
faiss-cpu==1.7.4
pypdf==4.0.1
python-dotenv==1.0.0
neo4j==5.14.1